Setup liveData in PopularMovies.kt, Move MovieModels.kt to model package, Change styles.xml theme to a material theme
