+++ app/src/main/java/com/practice_project/movieapp/model/MovieModels.kt
@@ -1,5 +1,5 @@
-package com.practice_project.movieapp.retrofit
+package com.practice_project.movieapp.model
 
-data class MovieList(val movies: List<Movie>)
+data class MovieList(val results: List<Movie>)
 
 data class Movie(val title: String, val id: Int, val release_date: String, val poster_path: String)
